+++ src/components/Grid/GridColumn.tsx
@@ -5,12 +5,12 @@ import { CLASSPREFIX as eccgui } from "../../configuration/constants";
 
 interface GridColumnProps extends ColumnDefaultProps {
     /**
-     * Column width is small, using 3 (or 2, on small viewports) parts out of 16.
+     * Column width is small, using 3 (or 2, on medium viewports) parts out of 16.
      * This boolean property is basically a quick switch for setting `{ md:2, lg:3 }`.
      */
     small?: boolean
     /**
-     * Column width is medium, using 5 (or 3, on small viewports) parts out of 16.
+     * Column width is medium, using 5 (or 3, on medium viewports) parts out of 16.
      * This boolean property is basically a quick switch for setting `{ md:3, lg:5 }`.
      */
     medium?: boolean
@@ -42,8 +42,8 @@ function GridColumn({
     if (medium) sizeConfig = { md:3, lg:5 };
     return (
         <CarbonColumn
-            {...otherProps}
             {...sizeConfig}
+            {...otherProps}
             className={
                 `${eccgui}-grid__column` +
                 (verticalAlign ? ` ${eccgui}-grid__column--vertical-` + verticalAlign : '') +