+++ src/components/Icon/IconButton.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import Button, { ButtonProps, AnchorOrButtonProps } from "../Button/Button";
 import {CLASSPREFIX as eccgui} from "../../configuration/constants";
 import Icon from "./Icon";
+import { TestIconProps } from "./TestIcon";
 import {ValidIconName} from "./canonicalIconNames";
 
 interface IconButtonProps extends Omit<ButtonProps, "icon" | "rightIcon" | "text" | "minimal" | "tooltip"> {
@@ -9,21 +10,21 @@ interface IconButtonProps extends Omit<ButtonProps, "icon" | "rightIcon" | "text
      * Canonical icon name, or an array of strings.
      * In case of the array the first valid icon name is used.
      */
-    name: ValidIconName | string[]
+    name: ValidIconName | string[] | React.ReactElement<TestIconProps>;
     /**
      * Button text, will be displayed as tooltip.
      */
-    text?: string,
+    text?: string;
     /**
      * If `text` should be set as HTML `title` attribute instead of attaching it as tooltip.
      * If true then `tooltipProps` is ignored.
      */
-    tooltipAsTitle?: boolean
+    tooltipAsTitle?: boolean;
     /**
      * Description for icon as accessibility fallback.
      * If not set then `text` is used.
      */
-    description?: string
+    description?: string;
     /**
      * Button is displayed with minimal styles (no borders, no background color).
      */
@@ -41,20 +42,23 @@ function IconButton({
     minimal=true,
     ...restProps
 }: IconButtonProps & AnchorOrButtonProps) {
+    const iconProps = {
+        small: restProps.small,
+        large: restProps.large,
+        tooltipText: tooltipAsTitle ? undefined : text,
+        tooltipProps: !!tooltipProps ? {hoverOpenDelay: 1000, ...tooltipProps} : {hoverOpenDelay: 1000},
+        description: description ? description : text,
+    };
+
     return (
         <Button
             title={tooltipAsTitle && text ? text : undefined}
             {...restProps}
-            icon={
-                <Icon
-                    name={name}
-                    small={restProps.small}
-                    large={restProps.large}
-                    tooltipText={tooltipAsTitle ? undefined : text}
-                    tooltipProps={!!tooltipProps ? {hoverOpenDelay: 1000, ...tooltipProps} : {hoverOpenDelay: 1000}}
-                    description={description ? description : text}
-                />
-            }
+            icon={(typeof name === "string" || Array.isArray(name)) ? (
+                <Icon name={name} {...iconProps} />
+            ) : (
+                React.cloneElement(name, iconProps)
+            )}
             className={`${eccgui}-button--icon ` + className}
             minimal={minimal}
         />