+++ src/components/Icon/stories/IconButton.stories.tsx
@@ -1,27 +1,34 @@
 import React from "react";
 import { ComponentStory, ComponentMeta } from "@storybook/react";
-
-import {
-    IconButton,
-} from "../../../index";
+import { LogoReact } from "@carbon/icons-react";
+import { IconButton,TestIcon } from "../../../index";
 import canonicalIcons from "./../canonicalIconNames";
 
 export default {
     title: "Components/IconButton",
     component: IconButton,
     argTypes: {
+        text: { control: "text" },
         name: {
             control: "select",
-            options: [...(Object.keys(canonicalIcons))],
+            options: [
+            "Test icon",
+                ...(Object.keys(canonicalIcons)),
+            ],
+            mapping: {
+            "Test icon": <TestIcon tryout={LogoReact} className="testclass-icon"/>,
+                ...(Object.keys(canonicalIcons)),
+            }
         },
     },
 } as ComponentMeta<typeof IconButton>;
 
 const Template: ComponentStory<typeof IconButton> = (args) => (
-    <IconButton {...args} text={args.name?.toString()}/>
+    <IconButton {...args} />
 );
 
 export const Default = Template.bind({});
 Default.args = {
-    name: "item-moremenu"
+    name: "item-moremenu",
+    text: "Tooltip text"
 }