+++ src/components/Tag/Tag.tsx
@@ -4,6 +4,13 @@ import {
     TagProps as BlueprintTagProps
 } from "@blueprintjs/core";
 import Color from "color";
+import Icon, { IconProps } from "../Icon/Icon";
+import { TestIconProps } from "../Icon/TestIcon";
+import { ValidIconName } from "../Icon/canonicalIconNames";
+import {
+    IntentTypes,
+    intentClassName
+} from "../../common/Intent";
 import decideContrastColorValue from "./../../common/utils/colorDecideContrastvalue";
 import { CLASSPREFIX as eccgui } from "../../configuration/constants";
 
@@ -12,12 +19,11 @@ export interface TagProps extends Omit<
     // currently not supported
     "active" |
     "fill" |
+    "icon" |
+    "intent" |
     "large" |
     "multiline" |
-    "intent" |
     "rightIcon"
-    // Removed round to have chip-like elements, since they are currently missing
-    // "round"
 > {
     // own properties
 
@@ -42,10 +48,19 @@ export interface TagProps extends Omit<
      * display tag in a large version
      */
     large?: boolean;
+    /**
+     * Meaning of the tag.
+     */
+    intent?: IntentTypes;
+    /**
+     * Icon displayed left from the tag label.
+     */
+    icon?: ValidIconName | React.ReactElement<IconProps> | React.ReactElement<TestIconProps>;
 
     // deprecated
 
     /**
+     * @deprecated
      * **deprecated**, use `minimal=false` plus `emphasis="stronger"`
      */
     emphasized?: never;
@@ -54,6 +69,8 @@ export interface TagProps extends Omit<
 function Tag({
     children,
     className = '',
+    intent,
+    icon,
     emphasis = "normal",
     minimal = true,
     small = false,
@@ -79,16 +96,19 @@ function Tag({
             }
         }
     }
+    const leftIcon = (!!icon && typeof icon === "string") ? <Icon name={icon} /> : icon;
     return (
         <BlueprintTag
             {...otherProps}
             className={
                 `${eccgui}-tag__item ${eccgui}-tag--${emphasis}emphasis` +
+                (!!intent ? ` ${intentClassName(intent)}` : '') +
                 (small ? ` ${eccgui}-tag--small` : '') +
                 (large ? ` ${eccgui}-tag--large` : '') +
                 (className ? ' ' + className : '')
             }
             minimal={minimal}
+            icon={!!leftIcon ? React.cloneElement(leftIcon, { small: !large}) : undefined}
         >
             {children}
         </BlueprintTag>