+++ src/components/index.ts
@@ -1,3 +1,6 @@
+export * from "./Badge/Badge";
+export * from "./Depiction/Depiction";
+export * from "./Icon";
 export * from "./ProgressBar";
 export * from "./Tag";
 export * from "./Tree/Tree";