+++ CHANGELOG.md
@@ -9,11 +9,33 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 * use option `--outputCss` for `yarn compile-scss` to get the transpiled CSS echoed out
 * `<ConfidenceValue/>` element: combines a value and a bar
+* `<Badge />` element to add more context like icons, text or numbers to another element
+* `<Depiction />` element to include different types of images controlling of resizing, ratio, shape
+* `<Tag />`: add support for `intent` property
+* `<ReactFlow />`:
+    * `scrollOnDrag`: Support to scroll the pane when going beyond the pane borders on all drag and connection operations.
+* `<MultiSelect />`:
+    * `requestDelay`: To delay requests on query changes and only fire the most recent request.
+* `<TestIcon />`: test icons without the need to define them via a canonical name before.
 
-### Change
+### Changed
 
 * upgrade to Carbon icons v11
 * switch from `carbon-components` to `@carbon/styles`
+* `<GridRow />` property `dontWrapColumns=true` only works for grids on medium sized and larger viewports
+
+### Fixed
+
+* `<WorkspaceContent />`: do not prevent wrapping the columns of the included grid
+
+### Deprecated
+
+* `<Grid />` property `fullWidth` is now deprecated as grids are always used for the full viewport width
+
+### Fixed
+
+* `<MultiSelect />`:
+    * Requests e.g. on slow networks could get mixed up, resulting in not showing the most recent results.
 
 ## [23.0.0] - 2022-11-18
 
@@ -54,7 +76,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 * fixed `<MultiSelect />` to correctly update created items that are selected while still maintaining a cache of all newly created items
 * do not change cursor to pointer by default on tooltip targets
 
-### Change
+### Changed
 
 * move style imports of CodeMirror layout to `extensions`
 * color configurations for react flow editor are not exported as modules anymore, they need to be fetched by `getColorConfiguration` method in JS directly